% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pet}
\alias{pet}
\title{A compilation of example mountain lakes in the US}
\format{
A dataframe with 19 rows and 5 variables:
\describe{
\item{name}{Name of the lake (character)}
\item{state}{US state abbreviation (character)}
\item{range}{The mountain range where the lake is located (character)}
\item{pct_mtn_area}{Percent of the mountain lake catchment that is mountain landform (numeric)}
\item{citation}{Citation where the lake is referred to as a mountain lake (character)}
}
}
\source{
{MountainLakes} A project to classify and report on the condition of mountain lakes in the US.
}
\usage{
pet
}
\description{
A dataframe containing lake names and location information for lakes that
have been identified as mountain lakes in published literature.
}
\keyword{datasets}
