% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{th_cond}
\alias{th_cond}
\title{Condition estimates for mountain and non-mountain lake populations at several thresholds}
\format{
A data frame with 675 rows and 15 variables:
\describe{
\item{Type}{The different mountain lake thresholds use for subpopulation estimates.
Follows the convention "MOUNTAIN_XXPCT" where XX is the percent threshold of mountain area
in the catchment required to be classified as a mountain lake (character)}
\item{Subpopulation}{Mountain lake (MTN_LAKE), non-mountain lake (NOT_MTN_LAKE), or all lakes (NATIONAL)}
\item{Indicator}{Condition indicator (character)}
\item{Category}{Category for the condition indicator (character)}
\item{nResp}{Number of sampled lakes in the indicator condition category (integer)}
\item{Estimate.P}{Estimated percent of subpopulation in the indicator condition category (percent)}
\item{StdError.P}{Standard error of percent of subpopulation in the indicator condition category (percent)}
\item{MarginofError.P}{Margin of error of percent of subpopulation in the indicator condition category (percent)}
\item{LCB95Pct.P}{Lower bound of 95\% confidence interval for percent of subpopulation in the indicator condition category (percent)}
\item{UCB95Pct.P}{Upper bound of 95\% confidence interval for percent of subpopulation in the indicator condition category (percent)}
\item{Estimate.U}{Estimated number of lakes of subpopulation in the indicator condition category (numeric)}
\item{StdError.U}{Standard error of number of lakes of subpopulation in the indicator condition category (numeric)}
\item{MarginofError.U}{Margin of error of number of lakes of subpopulation in the indicator condition category (numeric)}
\item{LCB95Pct.P}{Lower bound of 95\% confidence interval for the number of subpopulation in the indicator condition category (numeric)}
\item{UCB95Pct.P}{Upper bound of 95\% confidence interval for the number of subpopulation in the indicator condition category (numeric)}
}
}
\source{
{MountainLakes} A project to classify and report on the condition of mountain lakes in the US.
}
\usage{
th_cond
}
\description{
A dataset containing condition estimates for the National Lakes Assessment
target population with a subpopulation defined for mountain and non-mountain lakes.
Sampled lakes were classified as mountain lakes or non-mountain lakes based on the percent
of their LakeCat defined catchment that is mountain landform. Population weights are based
on a combination of the 2007, 2012, and 2017 lakes sampled in the National Lakes Assessment.
Produced with spsurvey package. See the IdentifyMountainLakes.Rmd script to learn how mountain
lakes are identified
}
\keyword{datasets}
